refactor: add DataManger class
